bug fix for read_maps() in case of 1D model
When only a 1D model is supplied, there is no inifieldfile and thus inifield_model = None. This would result in an error later in read_maps. The if-then statements ensures that for 1D models (with no fieldfile) empty lists [] are created such that the rest of the code still works